Inicio de sesión y creación de usuario con JWT, Django y react

+++ productos/models.py
@@ -3,8 +3,8 @@ from django.db import models
 class Producto(models.Model):
     nombre = models.CharField(max_length=100)
     descripcion = models.TextField()
-    costo = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
-    precio_venta = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    costo = models.IntegerField(null=True, blank=True)  # Cambiado a IntegerField
+    precio_venta = models.IntegerField(null=True, blank=True)  # Cambiado a IntegerField
     stock = models.IntegerField(null=True, blank=True)
 
     def __str__(self):
